samples/python/requirements.txt
deepgram-sdk>=6.0.0
pytest>=8.0.0
pytest-timeout>=2.0.0
pytest-asyncio>=0.23.0
